Move tool specs into core handlers
