+++ codex-rs/core/src/tools/handlers/apply_patch.lark
