+++ codex-rs/core/src/tools/handlers/goal/get_goal.rs
@@ -2,9 +2,9 @@ use crate::function_tool::FunctionCallError;
 use crate::tools::context::FunctionToolOutput;
 use crate::tools::context::ToolInvocation;
 use crate::tools::context::ToolPayload;
+use crate::tools::handlers::goal_spec::GET_GOAL_TOOL_NAME;
 use crate::tools::registry::ToolHandler;
 use crate::tools::registry::ToolKind;
-use codex_tools::GET_GOAL_TOOL_NAME;
 use codex_tools::ToolName;
 
 use super::CompletionBudgetReport;
