+++ codex-rs/core/src/tools/handlers/goal_spec.rs
@@ -3,9 +3,9 @@
 //! These specs expose goal read/update primitives to the model while keeping
 //! usage accounting system-managed.
 
-use crate::JsonSchema;
-use crate::ResponsesApiTool;
-use crate::ToolSpec;
+use codex_tools::JsonSchema;
+use codex_tools::ResponsesApiTool;
+use codex_tools::ToolSpec;
 use serde_json::json;
 use std::collections::BTreeMap;
