+++ codex-rs/core/src/tools/handlers/mod.rs
@@ -1,22 +1,34 @@
 pub(crate) mod agent_jobs;
+pub(crate) mod agent_jobs_spec;
 pub(crate) mod apply_patch;
+pub(crate) mod apply_patch_spec;
 mod dynamic;
 mod goal;
+pub(crate) mod goal_spec;
 mod mcp;
 mod mcp_resource;
+pub(crate) mod mcp_resource_spec;
 pub(crate) mod multi_agents;
 pub(crate) mod multi_agents_common;
+pub(crate) mod multi_agents_spec;
 pub(crate) mod multi_agents_v2;
 mod plan;
+pub(crate) mod plan_spec;
 mod request_permissions;
 mod request_plugin_install;
+pub(crate) mod request_plugin_install_spec;
 mod request_user_input;
+pub(crate) mod request_user_input_spec;
 mod shell;
+pub(crate) mod shell_spec;
 mod test_sync;
+pub(crate) mod test_sync_spec;
 mod tool_search;
+pub(crate) mod tool_search_spec;
 mod unavailable_tool;
 pub(crate) mod unified_exec;
 mod view_image;
+pub(crate) mod view_image_spec;
 
 use codex_sandboxing::policy_transforms::intersect_permission_profiles;
 use codex_sandboxing::policy_transforms::merge_permission_profiles;
