+++ codex-rs/core/src/tools/handlers/request_user_input_spec_tests.rs
@@ -1,8 +1,9 @@
 use super::*;
-use crate::JsonSchema;
 use codex_features::Feature;
 use codex_features::Features;
 use codex_protocol::config_types::ModeKind;
+use codex_tools::JsonSchema;
+use codex_tools::request_user_input_available_modes;
 use pretty_assertions::assert_eq;
 use std::collections::BTreeMap;
