+++ codex-rs/core/src/tools/handlers/shell_spec_tests.rs
