+++ codex-rs/core/src/tools/handlers/view_image_spec.rs
@@ -1,7 +1,7 @@
-use crate::JsonSchema;
-use crate::ResponsesApiTool;
-use crate::ToolSpec;
 use codex_protocol::models::VIEW_IMAGE_TOOL_NAME;
+use codex_tools::JsonSchema;
+use codex_tools::ResponsesApiTool;
+use codex_tools::ToolSpec;
 use serde_json::Value;
 use serde_json::json;
 use std::collections::BTreeMap;
