+++ codex-rs/core/src/tools/spec_plan.rs
@@ -1,72 +1,72 @@
-use crate::CommandToolOptions;
-use crate::REQUEST_PLUGIN_INSTALL_TOOL_NAME;
-use crate::REQUEST_USER_INPUT_TOOL_NAME;
-use crate::ResponsesApiNamespace;
-use crate::ResponsesApiNamespaceTool;
-use crate::ShellToolOptions;
-use crate::SpawnAgentToolOptions;
-use crate::TOOL_SEARCH_DEFAULT_LIMIT;
-use crate::TOOL_SEARCH_TOOL_NAME;
-use crate::ToolEnvironmentMode;
-use crate::ToolHandlerKind;
-use crate::ToolName;
-use crate::ToolRegistryPlan;
-use crate::ToolRegistryPlanParams;
-use crate::ToolSearchSource;
-use crate::ToolSearchSourceInfo;
-use crate::ToolSpec;
-use crate::ToolsConfig;
-use crate::ViewImageToolOptions;
-use crate::WebSearchToolOptions;
-use crate::coalesce_loadable_tool_specs;
-use crate::collect_code_mode_exec_prompt_tool_definitions;
-use crate::collect_request_plugin_install_entries;
-use crate::collect_tool_search_source_infos;
-use crate::create_apply_patch_freeform_tool;
-use crate::create_apply_patch_json_tool;
-use crate::create_close_agent_tool_v1;
-use crate::create_close_agent_tool_v2;
-use crate::create_code_mode_tool;
-use crate::create_create_goal_tool;
-use crate::create_followup_task_tool;
-use crate::create_get_goal_tool;
-use crate::create_image_generation_tool;
-use crate::create_list_agents_tool;
-use crate::create_list_mcp_resource_templates_tool;
-use crate::create_list_mcp_resources_tool;
-use crate::create_local_shell_tool;
-use crate::create_read_mcp_resource_tool;
-use crate::create_report_agent_job_result_tool;
-use crate::create_request_permissions_tool;
-use crate::create_request_plugin_install_tool;
-use crate::create_request_user_input_tool;
-use crate::create_resume_agent_tool;
-use crate::create_send_input_tool_v1;
-use crate::create_send_message_tool;
-use crate::create_shell_command_tool;
-use crate::create_shell_tool;
-use crate::create_spawn_agent_tool_v1;
-use crate::create_spawn_agent_tool_v2;
-use crate::create_spawn_agents_on_csv_tool;
-use crate::create_test_sync_tool;
-use crate::create_tool_search_tool;
-use crate::create_update_goal_tool;
-use crate::create_update_plan_tool;
-use crate::create_view_image_tool;
-use crate::create_wait_agent_tool_v1;
-use crate::create_wait_agent_tool_v2;
-use crate::create_wait_tool;
-use crate::create_web_search_tool;
-use crate::create_write_stdin_tool;
-use crate::default_namespace_description;
-use crate::dynamic_tool_to_loadable_tool_spec;
-use crate::local_tool::create_exec_command_tool_with_environment_id;
-use crate::mcp_tool_to_responses_api_tool;
-use crate::request_permissions_tool_description;
-use crate::request_user_input_tool_description;
-use crate::tool_registry_plan_types::agent_type_description;
+use crate::tools::code_mode::execute_spec::create_code_mode_tool;
+use crate::tools::code_mode::wait_spec::create_wait_tool;
+use crate::tools::handlers::agent_jobs_spec::create_report_agent_job_result_tool;
+use crate::tools::handlers::agent_jobs_spec::create_spawn_agents_on_csv_tool;
+use crate::tools::handlers::apply_patch_spec::create_apply_patch_freeform_tool;
+use crate::tools::handlers::apply_patch_spec::create_apply_patch_json_tool;
+use crate::tools::handlers::goal_spec::create_create_goal_tool;
+use crate::tools::handlers::goal_spec::create_get_goal_tool;
+use crate::tools::handlers::goal_spec::create_update_goal_tool;
+use crate::tools::handlers::mcp_resource_spec::create_list_mcp_resource_templates_tool;
+use crate::tools::handlers::mcp_resource_spec::create_list_mcp_resources_tool;
+use crate::tools::handlers::mcp_resource_spec::create_read_mcp_resource_tool;
+use crate::tools::handlers::multi_agents_spec::SpawnAgentToolOptions;
+use crate::tools::handlers::multi_agents_spec::create_close_agent_tool_v1;
+use crate::tools::handlers::multi_agents_spec::create_close_agent_tool_v2;
+use crate::tools::handlers::multi_agents_spec::create_followup_task_tool;
+use crate::tools::handlers::multi_agents_spec::create_list_agents_tool;
+use crate::tools::handlers::multi_agents_spec::create_resume_agent_tool;
+use crate::tools::handlers::multi_agents_spec::create_send_input_tool_v1;
+use crate::tools::handlers::multi_agents_spec::create_send_message_tool;
+use crate::tools::handlers::multi_agents_spec::create_spawn_agent_tool_v1;
+use crate::tools::handlers::multi_agents_spec::create_spawn_agent_tool_v2;
+use crate::tools::handlers::multi_agents_spec::create_wait_agent_tool_v1;
+use crate::tools::handlers::multi_agents_spec::create_wait_agent_tool_v2;
+use crate::tools::handlers::plan_spec::create_update_plan_tool;
+use crate::tools::handlers::request_plugin_install_spec::create_request_plugin_install_tool;
+use crate::tools::handlers::request_user_input_spec::REQUEST_USER_INPUT_TOOL_NAME;
+use crate::tools::handlers::request_user_input_spec::create_request_user_input_tool;
+use crate::tools::handlers::request_user_input_spec::request_user_input_tool_description;
+use crate::tools::handlers::shell_spec::CommandToolOptions;
+use crate::tools::handlers::shell_spec::ShellToolOptions;
+use crate::tools::handlers::shell_spec::create_exec_command_tool_with_environment_id;
+use crate::tools::handlers::shell_spec::create_local_shell_tool;
+use crate::tools::handlers::shell_spec::create_request_permissions_tool;
+use crate::tools::handlers::shell_spec::create_shell_command_tool;
+use crate::tools::handlers::shell_spec::create_shell_tool;
+use crate::tools::handlers::shell_spec::create_write_stdin_tool;
+use crate::tools::handlers::shell_spec::request_permissions_tool_description;
+use crate::tools::handlers::test_sync_spec::create_test_sync_tool;
+use crate::tools::handlers::tool_search_spec::create_tool_search_tool;
+use crate::tools::handlers::view_image_spec::ViewImageToolOptions;
+use crate::tools::handlers::view_image_spec::create_view_image_tool;
+use crate::tools::hosted_spec::WebSearchToolOptions;
+use crate::tools::hosted_spec::create_image_generation_tool;
+use crate::tools::hosted_spec::create_web_search_tool;
+use crate::tools::spec_plan_types::ToolHandlerKind;
+use crate::tools::spec_plan_types::ToolRegistryPlan;
+use crate::tools::spec_plan_types::ToolRegistryPlanParams;
+use crate::tools::spec_plan_types::agent_type_description;
 use codex_protocol::openai_models::ApplyPatchToolType;
 use codex_protocol::openai_models::ConfigShellToolType;
+use codex_tools::REQUEST_PLUGIN_INSTALL_TOOL_NAME;
+use codex_tools::ResponsesApiNamespace;
+use codex_tools::ResponsesApiNamespaceTool;
+use codex_tools::TOOL_SEARCH_DEFAULT_LIMIT;
+use codex_tools::TOOL_SEARCH_TOOL_NAME;
+use codex_tools::ToolEnvironmentMode;
+use codex_tools::ToolName;
+use codex_tools::ToolSearchSource;
+use codex_tools::ToolSearchSourceInfo;
+use codex_tools::ToolSpec;
+use codex_tools::ToolsConfig;
+use codex_tools::coalesce_loadable_tool_specs;
+use codex_tools::collect_code_mode_exec_prompt_tool_definitions;
+use codex_tools::collect_request_plugin_install_entries;
+use codex_tools::collect_tool_search_source_infos;
+use codex_tools::default_namespace_description;
+use codex_tools::dynamic_tool_to_loadable_tool_spec;
+use codex_tools::mcp_tool_to_responses_api_tool;
 use std::collections::BTreeMap;
 
 pub fn build_tool_registry_plan(
@@ -631,5 +631,5 @@ fn code_mode_namespace_name<'a>(
 }
 
 #[cfg(test)]
-#[path = "tool_registry_plan_tests.rs"]
+#[path = "spec_plan_tests.rs"]
 mod tests;
