+++ codex-rs/core/src/tools/spec_plan_tests.rs
@@ -1,27 +1,11 @@
 use super::*;
-use crate::AdditionalProperties;
-use crate::ConfiguredToolSpec;
-use crate::DiscoverablePluginInfo;
-use crate::DiscoverableTool;
-use crate::FreeformTool;
-use crate::JsonSchema;
-use crate::JsonSchemaPrimitiveType;
-use crate::JsonSchemaType;
-use crate::ResponsesApiNamespaceTool;
-use crate::ResponsesApiTool;
-use crate::ResponsesApiWebSearchFilters;
-use crate::ResponsesApiWebSearchUserLocation;
-use crate::ToolEnvironmentMode;
-use crate::ToolHandlerSpec;
-use crate::ToolName;
-use crate::ToolNamespace;
-use crate::ToolRegistryPlanDeferredTool;
-use crate::ToolRegistryPlanMcpTool;
-use crate::ToolsConfigParams;
-use crate::WaitAgentTimeoutOptions;
-use crate::create_exec_command_tool;
-use crate::mcp_call_tool_result_output_schema;
-use crate::request_user_input_available_modes;
+use crate::tools::handlers::multi_agents_spec::WaitAgentTimeoutOptions;
+use crate::tools::handlers::shell_spec::CommandToolOptions;
+use crate::tools::handlers::shell_spec::create_exec_command_tool;
+use crate::tools::spec_plan_types::ToolHandlerSpec;
+use crate::tools::spec_plan_types::ToolNamespace;
+use crate::tools::spec_plan_types::ToolRegistryPlanDeferredTool;
+use crate::tools::spec_plan_types::ToolRegistryPlanMcpTool;
 use codex_app_server_protocol::AppInfo;
 use codex_features::Feature;
 use codex_features::Features;
@@ -37,6 +21,23 @@ use codex_protocol::openai_models::ModelInfo;
 use codex_protocol::openai_models::WebSearchToolType;
 use codex_protocol::protocol::SessionSource;
 use codex_protocol::protocol::SubAgentSource;
+use codex_tools::AdditionalProperties;
+use codex_tools::ConfiguredToolSpec;
+use codex_tools::DiscoverablePluginInfo;
+use codex_tools::DiscoverableTool;
+use codex_tools::FreeformTool;
+use codex_tools::JsonSchema;
+use codex_tools::JsonSchemaPrimitiveType;
+use codex_tools::JsonSchemaType;
+use codex_tools::ResponsesApiNamespaceTool;
+use codex_tools::ResponsesApiTool;
+use codex_tools::ResponsesApiWebSearchFilters;
+use codex_tools::ResponsesApiWebSearchUserLocation;
+use codex_tools::ToolEnvironmentMode;
+use codex_tools::ToolName;
+use codex_tools::ToolsConfigParams;
+use codex_tools::mcp_call_tool_result_output_schema;
+use codex_tools::request_user_input_available_modes;
 use pretty_assertions::assert_eq;
 use serde_json::json;
 use std::collections::BTreeMap;
