+++ codex-rs/tools/src/lib.rs
@@ -1,63 +1,25 @@
 //! Shared tool definitions and Responses API tool primitives that can live
 //! outside `codex-core`.
 
-mod agent_job_tool;
-mod agent_tool;
-mod apply_patch_tool;
 mod code_mode;
 mod dynamic_tool;
-mod goal_tool;
 mod image_detail;
 mod json_schema;
-mod local_tool;
-mod mcp_resource_tool;
 mod mcp_tool;
-mod plan_tool;
 mod request_plugin_install;
-mod request_user_input_tool;
 mod responses_api;
 mod tool_config;
 mod tool_definition;
 mod tool_discovery;
-mod tool_registry_plan;
-mod tool_registry_plan_types;
 mod tool_spec;
-mod utility_tool;
-mod view_image;
 
-pub use agent_job_tool::create_report_agent_job_result_tool;
-pub use agent_job_tool::create_spawn_agents_on_csv_tool;
-pub use agent_tool::SpawnAgentToolOptions;
-pub use agent_tool::WaitAgentTimeoutOptions;
-pub use agent_tool::create_close_agent_tool_v1;
-pub use agent_tool::create_close_agent_tool_v2;
-pub use agent_tool::create_followup_task_tool;
-pub use agent_tool::create_list_agents_tool;
-pub use agent_tool::create_resume_agent_tool;
-pub use agent_tool::create_send_input_tool_v1;
-pub use agent_tool::create_send_message_tool;
-pub use agent_tool::create_spawn_agent_tool_v1;
-pub use agent_tool::create_spawn_agent_tool_v2;
-pub use agent_tool::create_wait_agent_tool_v1;
-pub use agent_tool::create_wait_agent_tool_v2;
-pub use apply_patch_tool::ApplyPatchToolArgs;
-pub use apply_patch_tool::create_apply_patch_freeform_tool;
-pub use apply_patch_tool::create_apply_patch_json_tool;
 pub use code_mode::augment_tool_spec_for_code_mode;
 pub use code_mode::code_mode_name_for_tool_name;
 pub use code_mode::collect_code_mode_exec_prompt_tool_definitions;
 pub use code_mode::collect_code_mode_tool_definitions;
-pub use code_mode::create_code_mode_tool;
-pub use code_mode::create_wait_tool;
 pub use code_mode::tool_spec_to_code_mode_tool_definition;
 pub use codex_protocol::ToolName;
 pub use dynamic_tool::parse_dynamic_tool;
-pub use goal_tool::CREATE_GOAL_TOOL_NAME;
-pub use goal_tool::GET_GOAL_TOOL_NAME;
-pub use goal_tool::UPDATE_GOAL_TOOL_NAME;
-pub use goal_tool::create_create_goal_tool;
-pub use goal_tool::create_get_goal_tool;
-pub use goal_tool::create_update_goal_tool;
 pub use image_detail::can_request_original_image_detail;
 pub use image_detail::normalize_output_image_detail;
 pub use image_detail::sanitize_original_image_detail;
@@ -66,20 +28,8 @@ pub use json_schema::JsonSchema;
 pub use json_schema::JsonSchemaPrimitiveType;
 pub use json_schema::JsonSchemaType;
 pub use json_schema::parse_tool_input_schema;
-pub use local_tool::CommandToolOptions;
-pub use local_tool::ShellToolOptions;
-pub use local_tool::create_exec_command_tool;
-pub use local_tool::create_request_permissions_tool;
-pub use local_tool::create_shell_command_tool;
-pub use local_tool::create_shell_tool;
-pub use local_tool::create_write_stdin_tool;
-pub use local_tool::request_permissions_tool_description;
-pub use mcp_resource_tool::create_list_mcp_resource_templates_tool;
-pub use mcp_resource_tool::create_list_mcp_resources_tool;
-pub use mcp_resource_tool::create_read_mcp_resource_tool;
 pub use mcp_tool::mcp_call_tool_result_output_schema;
 pub use mcp_tool::parse_mcp_tool;
-pub use plan_tool::create_update_plan_tool;
 pub use request_plugin_install::REQUEST_PLUGIN_INSTALL_APPROVAL_KIND_VALUE;
 pub use request_plugin_install::REQUEST_PLUGIN_INSTALL_PERSIST_ALWAYS_VALUE;
 pub use request_plugin_install::REQUEST_PLUGIN_INSTALL_PERSIST_KEY;
@@ -89,12 +39,6 @@ pub use request_plugin_install::RequestPluginInstallResult;
 pub use request_plugin_install::all_requested_connectors_picked_up;
 pub use request_plugin_install::build_request_plugin_install_elicitation_request;
 pub use request_plugin_install::verified_connector_install_completed;
-pub use request_user_input_tool::REQUEST_USER_INPUT_TOOL_NAME;
-pub use request_user_input_tool::create_request_user_input_tool;
-pub use request_user_input_tool::normalize_request_user_input_args;
-pub use request_user_input_tool::request_user_input_available_modes;
-pub use request_user_input_tool::request_user_input_tool_description;
-pub use request_user_input_tool::request_user_input_unavailable_message;
 pub use responses_api::FreeformTool;
 pub use responses_api::FreeformToolFormat;
 pub use responses_api::LoadableToolSpec;
@@ -102,7 +46,7 @@ pub use responses_api::ResponsesApiNamespace;
 pub use responses_api::ResponsesApiNamespaceTool;
 pub use responses_api::ResponsesApiTool;
 pub use responses_api::coalesce_loadable_tool_specs;
-pub(crate) use responses_api::default_namespace_description;
+pub use responses_api::default_namespace_description;
 pub use responses_api::dynamic_tool_to_loadable_tool_spec;
 pub use responses_api::dynamic_tool_to_responses_api_tool;
 pub use responses_api::mcp_tool_to_deferred_responses_api_tool;
@@ -115,6 +59,7 @@ pub use tool_config::ToolsConfig;
 pub use tool_config::ToolsConfigParams;
 pub use tool_config::UnifiedExecShellMode;
 pub use tool_config::ZshForkConfig;
+pub use tool_config::request_user_input_available_modes;
 pub use tool_definition::ToolDefinition;
 pub use tool_discovery::DiscoverablePluginInfo;
 pub use tool_discovery::DiscoverableTool;
@@ -129,27 +74,10 @@ pub use tool_discovery::ToolSearchSource;
 pub use tool_discovery::ToolSearchSourceInfo;
 pub use tool_discovery::collect_request_plugin_install_entries;
 pub use tool_discovery::collect_tool_search_source_infos;
-pub use tool_discovery::create_request_plugin_install_tool;
-pub use tool_discovery::create_tool_search_tool;
 pub use tool_discovery::filter_request_plugin_install_discoverable_tools_for_client;
 pub use tool_discovery::tool_search_result_source_to_loadable_tool_spec;
-pub use tool_registry_plan::build_tool_registry_plan;
-pub use tool_registry_plan_types::ToolHandlerKind;
-pub use tool_registry_plan_types::ToolHandlerSpec;
-pub use tool_registry_plan_types::ToolNamespace;
-pub use tool_registry_plan_types::ToolRegistryPlan;
-pub use tool_registry_plan_types::ToolRegistryPlanDeferredTool;
-pub use tool_registry_plan_types::ToolRegistryPlanMcpTool;
-pub use tool_registry_plan_types::ToolRegistryPlanParams;
 pub use tool_spec::ConfiguredToolSpec;
 pub use tool_spec::ResponsesApiWebSearchFilters;
 pub use tool_spec::ResponsesApiWebSearchUserLocation;
 pub use tool_spec::ToolSpec;
-pub use tool_spec::WebSearchToolOptions;
-pub use tool_spec::create_image_generation_tool;
-pub use tool_spec::create_local_shell_tool;
 pub use tool_spec::create_tools_json_for_responses_api;
-pub use tool_spec::create_web_search_tool;
-pub use utility_tool::create_test_sync_tool;
-pub use view_image::ViewImageToolOptions;
-pub use view_image::create_view_image_tool;
